Crate coinlist

Parses coin lists according to the coin list standard.

License

The coinlist crate is licensed under the Apache 2.0 License.

Structs

CoinExtensions

Extra information about a coin.

CoinInfo

Coin information.

CoinList

A list of coins.

TagDetails

Details about what a tag is.

Version

Semver version of the coin list.

Enums

ChainID

ID of a chain supporting coins.